Added unit tests for the fib problem.

+++ 001-fib/main.h
@@ -6,9 +6,9 @@
 #include <map>
 #include <cstdlib>
 
-unsigned long int fib_dp(int n);
-unsigned long int fib_dp_memo(int n);
-unsigned long int fib_dp_tab(int n);
-unsigned long int fib_reg(int n);
+unsigned long long int fib_dp(int n);
+unsigned long long int fib_dp_memo(int n);
+unsigned long long int fib_dp_tab(int n);
+unsigned long long int fib_reg(int n);
 
 #endif // MAIN_H
